7737 trigger event
