+++ apps/backend-functions/src/data/types.ts
@@ -18,6 +18,5 @@ export {
 } from '@blockframes/permissions/+state/permissions.firestore';
 export { PublicUser } from '@blockframes/user/+state/user.firestore';
 export { RequestDemoInformations } from '@blockframes/utils/request-demo';
-export { MovieAnalytics, MovieEventAnalytics } from '@blockframes/movie/+state/movie.firestore';
 export { NotificationTypes, NotificationDocument } from '@blockframes/notification/+state/notification.firestore';
 export { ScreeningEventDocument } from '@blockframes/event/+state/event.firestore';
