+++ libs/analytics/src/lib/components/movie-analytics-chart/movie-analytics-chart.component.ts
@@ -1,5 +1,5 @@
 import { Component, ChangeDetectionStrategy, Input } from '@angular/core';
-import { MovieAnalytics } from '@blockframes/movie/+state/movie.firestore';
+import { MovieAnalytics } from './movie-analytics.model';
 import { lineChartOptions } from './default-chart-options';
 
 const chartInfo = [
