+++ libs/analytics/src/lib/components/movie-analytics-chart/movie-analytics.model.ts
@@ -1,0 +1,32 @@
+const analyticsEvents = [
+  'removedFromWishlist',
+  'addedToWishlist',
+  'promoReelOpened',
+  'pageView',
+  'screeningRequested',
+  'askingPriceRequested'
+] as const;
+export type AnalyticsEvents = typeof analyticsEvents[number];
+
+export interface MovieEventAnalytics {
+  event_date: string,
+  event_name: AnalyticsEvents,
+  hits: number,
+  movieId: string
+}
+
+export interface MovieAnalytics {
+  id: string;
+  addedToWishlist: {
+    current: MovieEventAnalytics[],
+    past: MovieEventAnalytics[]
+  },
+  movieViews: {
+    current: MovieEventAnalytics[],
+    past: MovieEventAnalytics[]
+  },
+  promoReelOpened: {
+    current: MovieEventAnalytics[],
+    past: MovieEventAnalytics[]
+  }
+}
